bz18303: Ditch download of items froms shares on disconnect/quit.
Ditch downloading items from shares when we disconnect from share.
If the operation is user-initiated then ask for confirmation.

If it is out of our control then simply print an informational message
that this may have happened.

If we are quitting, disconnect from any active shares which will
cancel downloads.