Donation dialog fixes.
There were some problems with the donate dialog and this hopefully should
address those concerns.  In particular, the donate dialog can show even
when we have already asked the user 3 times at which point it is supposed
to wait 6 months before asking again.

I think the main problem is the donate_nothanks key: it keeps track of when
the user's said no.  But this is changed when the user clicked no and not
if they clicked on the X button at the window.

I've changed it so that:

(0) Clicking on the X is the same as clicking as no: the donate_nothanks
gets incremented and it is set in the preferences.

(1) clicking on yes sets the last_donate_time, which matches with what the
documentation says it is supposed to do.

(2) The show_donation dialog condition has been changed from
donate_counter == 0 and donate_nothanks < 3 to include
current time - last donate time < half year